Loading data to file in directory
Like with all the functions of this program, I am loading the results of
the functions into files (within the directory of the program). Problem:
some of the functions did not "return" but "print" . Had to change these
functions.
Problem: 3rd function (given_word) uses lists of lists. I don't like the
way this is returned (in between square brackets).